+++ es3/downloader.cpp
@@ -59,7 +59,7 @@ public:
 	{
 		return "writer"+int_to_string(get_class_limit());
 	}
-	virtual int get_class_limit() const
+	virtual size_t get_class_limit() const
 	{
 		return content_->ctx_->max_readers_;
 	}
